Forum complet

+++ src/main/java/com/example/akkar2/repository/LoanPapersRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface LoanPapersRepository  extends JpaRepository<LoanPapers, Integer> {
+public interface LoanPapersRepository extends JpaRepository<LoanPapers, Long> {
 }
